fix(hub): address drives-analytics review Important and Minor items
Lighten goal-alignment with a hist+dominant rail verdict instead of a
full window fetch; SQL AVG for mean pressures; funnel_scope honesty;
drop dead color branch; split I/O into drives_analytics_queries.py.
